Squashed 'praat/' changes from ddbdc78f..ee1e9000
ee1e9000 Merge pull request #571 from PaulBoersma/master
9c331908 header file bug for LineSpectralFrequencies
29b98be7 Merge pull request #570 from PaulBoersma/master
206023c7 for version 6.0.37
255254cd Merge pull request #569 from PaulBoersma/master
04e2cde5 date
278e765b bug fix \un trigraph
005171ff Merge pull request #567 from PaulBoersma/master
96a7fc95 Merge pull request #566 from davidweenink/master
84fa6bc6 SpeechSynthesizer: improvements for de, fi, fr, lv, it, hak, hr
0650d0f7 channels can be 0 again
d96d4df0 pairwise sum comments
5a866a33 Merge pull request #565 from davidweenink/master
61b20997 Make old commands work
fdefb56c Merge pull request #564 from PaulBoersma/master
5bb22212 Merge remote-tracking branch 'upstream/master'
7c9aa4af add 48k resampling script
af3b6625 2018
98fbdb72 modernize
1db92a2d rid some C89
4852d547 nets
2232ece6 Merge pull request #563 from davidweenink/master
04f87d8a Harmonize interface
b29dcd35 Harmonize interface
ce635498 Merge pull request #561 from davidweenink/master
a65d1c83 replace _and_ by _ between two selected objects in interace
9e63f16f replace _and_ by _ between two selected objects in interace
9142950f Merge pull request #560 from davidweenink/master
aabfa27a interface: downto
075c78d4 Merge pull request #559 from PaulBoersma/master
04092d60 long -> integer
b5c6d21d integer
9901017a Merge remote-tracking branch 'upstream/master'
b15619d7 Net messages
09971482 Merge pull request #558 from davidweenink/master
b7f0770f adapt script to dcomplex
0f586030 Merge pull request #557 from davidweenink/master
a863eeae NUMincompleteGammaFunction: return dcomplex value
255d76e1 Merge pull request #556 from davidweenink/master
9af5c246 gammaToneFilterResponseAtCentreFrequency: meaningfull names
2c8d46c8 Merge pull request #555 from davidweenink/master
6f47380b long -> integer
9f13692c SVD: read older versions
07808c60 regression integer <-> long
2b09b4a4 Cepstrum bugs
7f47f1f4 Merge pull request #554 from PaulBoersma/master
f5476a40 compatible FFNet_Pattern_Categories
16d04178 bugs from FFNet_Eigen
b69b9277 rid 60 bugs from FFNet
d68d685a Merge pull request #553 from PaulBoersma/master
48d926a0 got Covariance_getSignificanceOfOneVariance right
5771ccbe got Covariance_getSignificanceOfOneVariance() right
21129cc5 removed one bug from Covariance_getSignificanceOfOneVariance (but integer/double ndf mismatch still exists)
193fa62b Merge pull request #552 from PaulBoersma/master
9224bf9f Collection_extensions bugs
99e4ba81 typos
6f115763 Merge pull request #551 from PaulBoersma/master
5b946164 made LongSound_extensions compatible with Windows
35a1c643 TextGrid_extensions bugs
d43466aa EspeakVoice bug
9d4b4e5d Sound_extensions compilation bugs
3bb45539 Procrustes Windows bugs
1be8d8a8 Polygon_extensions bugs
9b82141e Matrix_extensions compilable on Windows
b5b069ab ClassificationTable compatibility bugs
d87e75b8 more AffineTransform bugs
8563d605 eleven Tube compatibility bugs
3ce3497c Procrustes compilation bug
98752231 Spectrogram_extensions Windows type bugs
064f267d rid Spectrum_extensions type bugs
9bc89737 rid many DataModeler Windows compatibility bugs
8ce056c7 rid AffineTransform compilation bug
b4fa6249 Merge pull request #550 from PaulBoersma/master
c1718778 rid Polynomial compilation bugs
4fa6b1d2 virtual layers 2
96c6c4d6 virtual layers 1
acf11a60 Merge pull request #549 from PaulBoersma/master
09d707b0 Windows compat
1906dc46 rid SimpleInt
c7b36bd3 rid fcomplex
7d9080bf Net makefiles
d8aeeca1 Net
99ac3ab2 Merge pull request #548 from PaulBoersma/master
de29ad8c Melder_isWhiteSpace
b51f2fef Merge pull request #547 from davidweenink/master
17d639d6 SVD manual pages
d24e11c7 Merge pull request #546 from davidweenink/master
1dfb1bc9 test_gsl.praat rewrite
f7956b3c SVD_def
d398bbcb Merge pull request #545 from davidweenink/master
5c80f637 NUM2.cpp remove not used procedure
bdb223ab + test_SVD.praat
eada1377 Merge pull request #544 from davidweenink/master
91a9f458 require
2ee0cb14 Merge https://github.com/praat/praat into svd
d42fb3bb Merge pull request #543 from PaulBoersma/master
befe62df comments
3a298a90 integrate channel muting
155a3da0 iswspace
cc4a12b6 SVD repare number of rows smaller than number of columns
60ada73a must -> should
ffd8bdcf Revert "automatically: must -> should"
76f1453e automatically: must -> should
0ed92af9 must -> should
1e2ba602 Merge pull request #541 from davidweenink/master
7088d8ac test_Ltas_reportSpectralTilt.praat don't fail on random data (2)
832bb6d9 typo
236387fc test_Ltas_reportSpectralTilt.praat don't fail on random data
9ef9886b espeak_io.cpp remove doubly included header file
0f0aa6ac Merge pull request #540 from davidweenink/master
310075ac test_Ltas_reportSpectralTilt.praat don't fail
11256d48 bool, integer, require
e6962cf3 require, integer
c7949b82 Merge pull request #539 from davidweenink/master
87568632 modify espeak data for big-endian machines
5e811513 Revert "temporary fix for big-endian machines"
48305e43 Merge https://github.com/praat/praat
dfdd4262 temporary fix for big-endian machines
d5b8294a correct bug in Chinese pronunciation
dadab122 require, integer
5e7a8c7b Merge pull request #538 from PaulBoersma/master
2b7419eb rid reset() for autonumvec and autonummat
36665bf8 pre-incremented enums
52931b87 fewer void pointers
e489d5af Merge pull request #537 from davidweenink/master
a9a965a6 require, integer
10be866e prompted tables in ManPage HTML files
5ab1a66c rid some void pointers
b76c6a85 Merge pull request #536 from davidweenink/master
5245dabc require, integer
ce090539 require, integer
70a43e58 require, integer
0a3e5254 standardize layout
458de434 More bools and integers
8f165011 FileInMemory more use of require
d254a587 Permutation more tests
f0ae9545 NUMtensor3
9bedd8ac Merge pull request #535 from davidweenink/master
0322d0ed FileInMemoryManager clean up
4f5219b8 Merge pull request #534 from davidweenink/master
c2f65e2d Melder_require and integer
86db7c4b espeak-ng simplified procedure to add changes from upstream in praat
f83b5952 allow LongSound editors to receive spurious v_dataChanged messages
16c53dda set default number of periods for Sound To Harmonicity (cc) back to 1.0 instead of 4.5
22c15edd Merge pull request #533 from davidweenink/master
21dc0d44 Soud_draw_where extra test
48c5fab2 code cleanup
8a5428e0 Merge pull request #531 from davidweenink/master
38ee932e Default colour is Black
88523d4f Sound_extensions: integers and bools
05423299 fix merge conflict
ea9ced06 Sound_and_Pitch_to_Spectrogram domain
bc723aee Merge pull request #530 from PaulBoersma/master
56e4c3fc rid some casts
f61f8504 Merge pull request #529 from PaulBoersma/master
32d46ce2 inplace
45224a2c Language updates for lv, it, ar and my
c914c238 objects listing more alhabetical
188afa34 test FileInMemoryManager IO
729cc8be Merge pull request #528 from PaulBoersma/master
35ce31d7 Melder_iround
221c7410 Melder_iceiling
b0afe5a9 Merge https://github.com/praat/praat
0717c922 Melder_ifloor
76043a1b Update from espeak/master
469fda03 NUM2.h reordering comments

git-subtree-dir: praat
git-subtree-split: ee1e9000d527c9040e2aed42cd081115dbfe6d90

+++ dwsys/DoublyLinkedList.cpp
@@ -1,6 +1,6 @@
 /* DoublyLinkedList.cpp
  *
- * Copyright (C) 2011-2013, 2015-2016 David Weenink
+ * Copyright (C) 2011-2017 David Weenink
  *
  * This code is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -73,7 +73,7 @@ void DoublyLinkedList_addFront (DoublyLinkedList me, DoublyLinkedNode node) {
 		my back = node;
 		node -> next = nullptr;
 		node -> prev = nullptr;
-		my numberOfNodes++;
+		my numberOfNodes ++;
 	}
 }
 
@@ -94,7 +94,7 @@ void DoublyLinkedList_addBefore (DoublyLinkedList me, DoublyLinkedNode pos, Doub
 		pos -> prev -> next = node;
 	}
 	pos -> prev = node;
-	my numberOfNodes++;
+	my numberOfNodes ++;
 }
 
 void DoublyLinkedList_addAfter (DoublyLinkedList me, DoublyLinkedNode pos, DoublyLinkedNode node) {
@@ -106,7 +106,7 @@ void DoublyLinkedList_addAfter (DoublyLinkedList me, DoublyLinkedNode pos, Doubl
 		pos -> next -> prev = node;
 	}
 	pos -> next = node;
-	my numberOfNodes++;
+	my numberOfNodes ++;
 }
 
 void DoublyLinkedList_remove (DoublyLinkedList me, DoublyLinkedNode node) {
@@ -124,11 +124,11 @@ void DoublyLinkedList_remove (DoublyLinkedList me, DoublyLinkedNode node) {
 		node -> next -> prev = node -> prev;
 	}
 	forget (node);
-	my numberOfNodes--;
+	my numberOfNodes --;
 }
 
 // Preconditions:
-//	from and to must be part of the list
+//	from and to should be part of the list
 //	from must occur before to
 void DoublyLinkedList_sortPart (DoublyLinkedList me, DoublyLinkedNode from, DoublyLinkedNode to) {
 	// Save data
@@ -163,21 +163,21 @@ void DoublyLinkedList_sortPart (DoublyLinkedList me, DoublyLinkedNode from, Doub
 
 void DoublyLinkedList_sort (DoublyLinkedList me) {
 	Data_CompareHook::FunctionType compare = my v_getCompareHook ().get();
-	long increment = 1;
+	integer increment = 1;
 	DoublyLinkedNode front = my front, back;
 	for (;;) {
 		DoublyLinkedNode node1 = front;
 		front = nullptr;
 		back = nullptr;
 
-		long numberOfMerges = 0;
+		integer numberOfMerges = 0;
 
 		while (node1) {
 			DoublyLinkedNode node2 = node1, node;
-			long node1size = 0;
+			integer node1size = 0;
 			numberOfMerges++;
 
-			for (long i = 1; i <= increment; i++) {
+			for (integer i = 1; i <= increment; i++) {
 				node1size++;
 				node2 = node2 -> next;
 				if (! node2) {
@@ -185,7 +185,7 @@ void DoublyLinkedList_sort (DoublyLinkedList me) {
 				}
 			}
 
-			long node2size = increment;
+			integer node2size = increment;
 
 			while (node1size > 0 || (node2size > 0 && node2)) { // merge node1 and node2
 				if (node1size == 0) {
@@ -219,4 +219,4 @@ void DoublyLinkedList_sort (DoublyLinkedList me) {
 	my back = back;
 }
 
-// end of file DoublyLinkedList.cpp
+/* end of file DoublyLinkedList.cpp */
